# lib/api-zod/src/generated/types/confirmShamingResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guardian Agent API — dark pattern detection and checkout protection
 * OpenAPI spec version: 0.1.0
 */

export interface ConfirmShamingResult {
  detected: boolean;
  shamingText: string;
  rewrittenText: string;
}