# lib/api-zod/src/generated/types/createReportBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guardian Agent API — dark pattern detection and checkout protection
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReportBody {
  domain: string;
  url: string;
  trustScore: number;
  falseUrgencyDetected: boolean;
  falseScarcityDetected: boolean;
  confirmShamingDetected: boolean;
  hiddenFeesDetected: boolean;
  preCheckedAddOnsDetected: boolean;
  misdirectionDetected: boolean;
  totalPatternsDetected: number;
  /** @nullable */
  hiddenFeesTotal?: number | null;
  summary: string;
}