# lib/api-zod/src/generated/types/detectDarkPatternsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guardian Agent API — dark pattern detection and checkout protection
 * OpenAPI spec version: 0.1.0
 */
import type { FormFieldInfo } from "./formFieldInfo";

export interface DetectDarkPatternsBody {
  /** The website domain being analyzed */
  domain: string;
  /** Extracted text content from the page */
  pageText: string;
  /** Text from countdown timer elements found on the page */
  timerElements?: string[];
  /** Text from stock/scarcity alert elements */
  stockAlerts?: string[];
  /** Text from all buttons on the page */
  buttonLabels?: string[];
  /** Form field info including pre-checked state */
  formFields?: FormFieldInfo[];
  /** All price strings found on the page */
  priceStrings?: string[];
}