# lib/api-zod/src/generated/types/domainOffenderSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guardian Agent API — dark pattern detection and checkout protection
 * OpenAPI spec version: 0.1.0
 */

export interface DomainOffenderSummary {
  domain: string;
  totalPatternsDetected: number;
  trustScore: number;
  tier: string;
  totalScans: number;
}