# lib/api-zod/src/generated/types/formFieldInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guardian Agent API — dark pattern detection and checkout protection
 * OpenAPI spec version: 0.1.0
 */

export interface FormFieldInfo {
  id: string;
  label: string;
  type: string;
  /** @nullable */
  checked?: boolean | null;
  /** @nullable */
  optional?: boolean | null;
}