# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guardian Agent API — dark pattern detection and checkout protection
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  totalScans: number;
  totalPatternsDetected: number;
  /** Estimated total money saved in dollars */
  totalHiddenFeesBlocked: number;
  totalDomainsTracked: number;
  totalReportsLast30Days: number;
  avgTrustScore: number;
  goldTierDomains: number;
  highManipulationDomains: number;
}